# docs/api/swagger.yaml
definitions:
...
  github_com_itering_subscan_plugins_evm_dao.AccountTokenJson:
    properties:
      balance:
        type: number
      category:
        type: string
      contract:
        type: string
      decimals:
        type: integer
      name:
        type: string
      symbol:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.AccountsJson:
    properties:
      balance:
        type: number
      evm_account:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.Contract:
    properties:
      CompileSettings:
        items:
          type: integer
        type: array
      abi:
        items:
          type: integer
        type: array
      address:
        type: string
      block_num:
        type: integer
      compiler_version:
        type: string
      constructor_arguments:
        type: string
      contract_name:
        type: string
      creation_bytecode:
        type: string
      creation_code:
        type: string
      deploy_at:
        type: integer
      deploy_code_hash:
        type: string
      deployer:
        type: string
      eip_standard:
        type: string
      event_identifiers:
        items:
          type: integer
        type: array
      evm_version:
        type: string
      external_libraries:
        items:
          type: integer
        type: array
      extrinsic_index:
        type: string
      method_identifiers:
        items:
          type: integer
        type: array
      optimization_runs:
        type: integer
      optimize:
        type: boolean
      precompile:
        type: integer
      proxy_implementation:
        type: string
      source_code:
        type: string
      transaction_count:
        type: integer
      tx_hash:
        type: string
      verify_status:
        type: string
      verify_time:
        type: integer
      verify_type:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.ContractsJson:
    properties:
      address:
        type: string
      contract_name:
        type: string
      transaction_count:
        type: integer
      verify_status:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.Erc721Holders:
    properties:
      contract:
        type: string
      holder:
        type: string
      metadata:
        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.Metadata'
      storage_url:
        type: string
      token_id:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.EvmBlock:
    properties:
      author:
        type: string
      base_fee_per_gas:
        type: number
      block_hash:
        type: string
      block_num:
        type: integer
      block_size:
        type: number
      difficulty:
        type: number
      extra_data:
        type: string
      gas_limit:
        type: number
      gas_used:
        type: number
      logs_bloom:
        type: string
      miner:
        type: string
      parent_hash:
        type: string
      receipts_root:
        type: string
      seal_fields:
        type: string
      sha3_uncles:
        type: string
      state_root:
        type: string
      timestamp:
        type: integer
      total_difficulty:
        type: number
      transaction_count:
        type: integer
      transactions_root:
        type: string
      uncles:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.EvmBlockJson:
    properties:
      block_num:
        type: integer
      block_timestamp:
        type: integer
      miner:
        type: string
      transactions:
        type: integer
    type: object
  github_com_itering_subscan_plugins_evm_dao.Metadata:
    properties:
      animation_url:
        type: string
      attributes:
        items:
          additionalProperties: true
          type: object
        type: array
      background_color:
        type: string
      description:
        type: string
      external_url:
        type: string
      image:
        type: string
      name:
        type: string
      youtube_url:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.Token:
    properties:
      base_token_uri:
        type: string
      category:
        type: string
      contract:
        type: string
      decimals:
        type: integer
      holders:
        type: integer
      name:
        type: string
      symbol:
        type: string
      totalSupply:
        type: number
      transfer_count:
        type: integer
    type: object
  github_com_itering_subscan_plugins_evm_dao.TokenHolder:
    properties:
      balance:
        type: number
      contract:
        type: string
      holder:
        type: string
    type: object
  github_com_itering_subscan_plugins_evm_dao.TokenTransferJson:
    properties:
      category:
        type: string
      contract:
        type: string
      create_at:
        type: integer
      decimals:
        type: integer
      from:
        type: string
      hash:
        type: string
      id:
        type: integer
      name:
        type: string
      storage_url:
        type: string
      symbol:
        type: string
      to:
        type: string
      token_id:
        type: string
      value:
        type: number
    type: object
  github_com_itering_subscan_plugins_evm_dao.Transaction:
    properties:
      block_num:
        type: integer
      block_timestamp:
        type: integer
      contract:
        type: string
      cumulative_gas_used:
        type: number
      effective_gas_price:
        description: eip 1559
        type: number
      extrinsic_index:
        type: string
      from_address:
        type: string
      gas_limit:
        type: number
      gas_price:
        type: number
      gas_used:
        type: number
      hash:
        type: string
      input_data:
        type: string
      max_fee_per_gas:
        type: number
      max_priority_fee_per_gas:
        type: number
      nonce:
        type: integer
      precompile:
        type: integer
      r:
        type: string
      s:
        type: string
      success:
        type: boolean
      to_address:
        type: string
      transaction_id:
        description: pk
        type: integer
      transaction_index:
        type: integer
      txn_type:
        type: integer
      v:
        type: integer
      value:
        type: number
    type: object
  github_com_itering_subscan_plugins_evm_dao.TransactionSampleJson:
    properties:
      block_num:
        type: integer
      block_timestamp:
        type: integer
      create:
        type: string
      from_address:
        type: string
      hash:
        type: string
      to_address:
        type: string
      transaction_id:
        type: integer
      value:
        type: number
    type: object
...
  plugins_evm_http.EvmAccountParams:
    properties:
      address:
        type: string
      after:
        minLength: 0
        type: string
      before:
        minLength: 0
        type: string
      include_contracts:
        type: boolean
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  plugins_evm_http.EvmBlockParams:
    properties:
      block_num:
        minimum: 0
        type: integer
      hash:
        type: string
    type: object
  plugins_evm_http.EvmBlocks:
    properties:
      after:
        minimum: 0
        type: integer
      before:
        minimum: 0
        type: integer
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  plugins_evm_http.EvmContractSolcVersionsParam:
    properties:
      releases:
        type: boolean
    type: object
  plugins_evm_http.J:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      ttl:
        type: integer
    type: object
  plugins_evm_http.accountTokensParams:
    properties:
      address:
        type: string
      category:
        enum:
        - erc20
        - erc721
        type: string
    required:
    - address
    type: object
  plugins_evm_http.collectiblesParams:
    properties:
      address:
        type: string
      after:
        minLength: 0
        type: string
      before:
        minLength: 0
        type: string
      contract:
        type: string
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  plugins_evm_http.contractParams:
    properties:
      address:
        type: string
    required:
    - address
    type: object
  plugins_evm_http.contractsParams:
    properties:
      after:
        minLength: 0
        type: string
      before:
        minLength: 0
        type: string
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  plugins_evm_http.tokenHolderParams:
    properties:
      after:
        minLength: 0
        type: string
      before:
        minLength: 0
        type: string
      row:
        maximum: 100
        minimum: 1
        type: integer
      token_address:
        type: string
    required:
    - token_address
    type: object
  plugins_evm_http.tokenListParams:
    properties:
      after:
        minLength: 0
        type: string
      before:
        minLength: 0
        type: string
      category:
        enum:
        - erc20
        - erc721
        type: string
      contract:
        type: string
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
  plugins_evm_http.tokenTransferParams:
    properties:
      address:
        type: string
      after:
        minimum: 0
        type: integer
      before:
        minimum: 0
        type: integer
      category:
        enum:
        - erc20
        - erc721
        type: string
      row:
        maximum: 100
        minimum: 1
        type: integer
      token_address:
        type: string
    type: object
  plugins_evm_http.transactionParam:
    properties:
      hash:
        type: string
    required:
    - hash
    type: object
  plugins_evm_http.transactionsParams:
    properties:
      address:
        type: string
      after:
        minimum: 0
        type: integer
      before:
        minimum: 0
        type: integer
      block_num:
        minimum: 0
        type: integer
      row:
        maximum: 100
        minimum: 1
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/plugin/evm/account/tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.accountTokensParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.AccountTokenJson'
                  type: array
              type: object
      summary: Get account tokens
      tags:
      - EVM
  /api/plugin/evm/accounts:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.EvmAccountParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    lixst:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.AccountsJson'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm accounts list
      tags:
      - EVM
  /api/plugin/evm/block:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.EvmBlockParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.EvmBlock'
              type: object
      summary: Evm block info
      tags:
      - EVM
  /api/plugin/evm/blocks:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.EvmBlocks'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.EvmBlockJson'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm blocks
      tags:
      - EVM
  /api/plugin/evm/contract:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.contractParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.Contract'
              type: object
      summary: Evm contract info
      tags:
      - EVM
...
  /api/plugin/evm/contracts:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.contractsParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.ContractsJson'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm contract list
      tags:
      - EVM
  /api/plugin/evm/token/erc721/collectibles:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.collectiblesParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.Erc721Holders'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm Erc721 collectibles
      tags:
      - EVM
  /api/plugin/evm/token/holder:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.tokenHolderParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    holders:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.TokenHolder'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm token holder
      tags:
      - EVM
  /api/plugin/evm/token/transfer:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.tokenTransferParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    pagination:
                      type: object
                    transfers:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.TokenTransferJson'
                      type: array
                  type: object
              type: object
      summary: Evm token transfer
      tags:
      - EVM
  /api/plugin/evm/tokens:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.tokenListParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.Token'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm token list
      tags:
      - EVM
  /api/plugin/evm/transaction:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.transactionParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.Transaction'
              type: object
      summary: Evm transaction info
      tags:
      - EVM
  /api/plugin/evm/transactions:
    post:
      consumes:
      - application/json
      parameters:
      - description: params
        in: body
        name: params
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.transactionsParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  properties:
                    list:
                      items:
                        $ref: '#/definitions/github_com_itering_subscan_plugins_evm_dao.TransactionSampleJson'
                      type: array
                    pagination:
                      type: object
                  type: object
              type: object
      summary: Evm transactions
      tags:
      - EVM
...
  /api/scan/evm/contract/solcs:
    post:
      consumes:
      - application/json
      parameters:
      - description: param
        in: body
        name: param
        required: true
        schema:
          $ref: '#/definitions/plugins_evm_http.EvmContractSolcVersionsParam'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/plugins_evm_http.J'
            - properties:
                data:
                  items:
                    type: string
                  type: array
              type: object
      summary: EVM contract solc versions
      tags:
      - EVM
...
swagger: "2.0"